ODE Solver with Automatic Differentiation, in C++ Header Files

odelia provides an ODE solver in C++ header files, using an adaptive-step Runge-Kutta 4-5 method, with an interface to R via Rcpp. ODE systems can be templated on their scalar type to support automatic differentiation, enabling exact gradients for parameter fitting.

Details

The DESCRIPTION file: This package was not yet installed at build time.

Package Content

Index: This package was not yet installed at build time.

Author

Daniel Falster [aut, cre] (ORCID: <https://orcid.org/0000-0002-9814-092X>), Richard FitzJohn [aut], Andrew O'Reilly-Nugent [aut]

Maintainer

Daniel Falster <daniel.falster@unsw.edu.au>

References

http://www.boost.org/